backend/requirements.txt
djangorestframework 
django-cors-headers
groq
python-dotenv
openai
reportlab